refactor: combine top rated, popular & upcoming sections local data sources & daos with movies & tv shows
