+++ localDatasource/src/main/java/com/amsterdam/localdatasource/roomDataBase/datasource/MovieLocalDataSourceImpl.kt
@@ -6,9 +6,13 @@ import com.amsterdam.localdatasource.roomDataBase.daos.MovieDao
 import com.amsterdam.repository.datasource.local.MovieLocalSource
 import com.amsterdam.repository.dto.local.LocalMovieDto
 import com.amsterdam.repository.dto.local.MovieCategoryCrossRefDto
+import com.amsterdam.repository.dto.local.PopularMovieDto
 import com.amsterdam.repository.dto.local.SearchMovieCrossRefDto
+import com.amsterdam.repository.dto.local.TopRatedMovieDto
+import com.amsterdam.repository.dto.local.UpcomingMovieDto
 import com.amsterdam.repository.dto.local.relation.MovieWithCategories
 import com.amsterdam.repository.dto.local.utils.SearchType
+import kotlinx.datetime.Instant
 import javax.inject.Inject
 
 
@@ -90,6 +94,64 @@ class MovieLocalDataSourceImpl @Inject constructor(
 
     override suspend fun getTopRatedMovies(storedLanguage: String): List<LocalMovieDto> {
         return movieDao.getTopRatedMovies(storedLanguage)
+    }
 
+    @Transaction
+    override suspend fun addPopularMovies(movies: List<LocalMovieDto>) {
+        movieDao.insertMovies(movies)
+        val entries = movies.map { movie ->
+            PopularMovieDto(
+                movieId = movie.movieId,
+                storedLanguage = movie.storedLanguage,
+                dateAdded = movie.insertedDate
+            )
+        }
+        movieDao.insertPopularMovies(entries)
+    }
+
+    override suspend fun deleteExpiredPopularMovies(
+        expirationTime: Instant,
+        storedLanguage: String
+    ) {
+        movieDao.deleteExpiredPopularMovies(expirationTime, storedLanguage)
+    }
+
+    @Transaction
+    override suspend fun addTopRatedMovies(movies: List<LocalMovieDto>) {
+        movieDao.insertMovies(movies)
+        val entries = movies.map { movie ->
+            TopRatedMovieDto(
+                movieId = movie.movieId,
+                storedLanguage = movie.storedLanguage
+            )
+        }
+        movieDao.insertTopRatedMovies(entries)
+    }
+
+    override suspend fun deleteAllExpiredTopRatedMovies(
+        expirationTime: Instant,
+        storedLanguage: String
+    ) {
+        movieDao.deleteAllExpiredTopRatedMovies(expirationTime, storedLanguage)
+    }
+
+    @Transaction
+    override suspend fun addUpcomingMovies(movies: List<LocalMovieDto>) {
+        movieDao.insertMovies(movies)
+        val entries = movies.map { movie ->
+            UpcomingMovieDto(
+                movieId = movie.movieId,
+                storedLanguage = movie.storedLanguage,
+                dateAdded = movie.insertedDate
+            )
+        }
+        movieDao.insertUpcomingMovies(entries)
+    }
+
+    override suspend fun deleteExpiredUpcomingMovies(
+        expirationTime: Instant,
+        storedLanguage: String
+    ) {
+        movieDao.deleteExpiredUpcomingMovies(expirationTime, storedLanguage)
     }
 }
