--- localDatasource/src/main/java/com/amsterdam/localdatasource/roomDataBase/datasource/PopularMovieLocalDataSourceImpl.kt
@@ -1,36 +1,0 @@
-package com.amsterdam.localdatasource.roomDataBase.datasource
-
-import androidx.room.Transaction
-import com.amsterdam.localdatasource.roomDataBase.daos.MovieDao
-import com.amsterdam.localdatasource.roomDataBase.daos.PopularMovieDao
-import com.amsterdam.repository.datasource.local.PopularMovieLocalSource
-import com.amsterdam.repository.dto.local.LocalMovieDto
-import com.amsterdam.repository.dto.local.PopularMovieDto
-import kotlinx.datetime.Instant
-import javax.inject.Inject
-
-
-class PopularMovieLocalDataSourceImpl @Inject constructor(
-    private val movieDao: MovieDao,
-    private val popularMovieDao: PopularMovieDao
-) : PopularMovieLocalSource {
-    @Transaction
-    override suspend fun addPopularMovies(movies: List<LocalMovieDto>) {
-        movieDao.insertMovies(movies)
-        val entries = movies.map { movie ->
-            PopularMovieDto(
-                movieId = movie.movieId,
-                storedLanguage = movie.storedLanguage,
-                dateAdded = movie.insertedDate
-            )
-        }
-        popularMovieDao.insertPopularMovies(entries)
-    }
-
-    override suspend fun deleteExpiredPopularMovies(
-        expirationTime: Instant,
-        storedLanguage: String
-    ) {
-        popularMovieDao.deleteExpiredPopularMovies(expirationTime, storedLanguage)
-    }
-}
