+++ repository/src/main/java/com/amsterdam/repository/datasource/local/TvShowLocalSource.kt
@@ -2,6 +2,7 @@ package com.amsterdam.repository.datasource.local
 
 import com.amsterdam.repository.dto.local.LocalTvShowDto
 import com.amsterdam.repository.dto.local.relation.TvShowWithCategory
+import kotlinx.datetime.Instant
 
 interface TvShowLocalSource {
     suspend fun getTvShowsBySearchKeywordSortedByInterest(
@@ -30,5 +31,14 @@ interface TvShowLocalSource {
     suspend fun getTvShowById(tvShowId: Long, storedLanguage: String): LocalTvShowDto?
 
     suspend fun getPopularTvShows(storedLanguage: String): List<TvShowWithCategory>
+
     suspend fun getTopRatedTvShows(storedLanguage: String): List<LocalTvShowDto>
+
+    suspend fun addPopularTvShows(tvShows: List<LocalTvShowDto>)
+
+    suspend fun deleteExpiredPopularTvShows(expirationTime: Instant, storedLanguage: String)
+
+    suspend fun addTopRatedTvShows(tvShows: List<LocalTvShowDto>)
+
+    suspend fun deleteExpiredTopRatedTvShows(expirationTime: Instant, storedLanguage: String)
 }
