--- repository/src/main/java/com/amsterdam/repository/datasource/local/UpcomingMovieLocalSource.kt
@@ -1,9 +1,0 @@
-package com.amsterdam.repository.datasource.local
-
-import com.amsterdam.repository.dto.local.LocalMovieDto
-import kotlinx.datetime.Instant
-
-interface UpcomingMovieLocalSource {
-    suspend fun addUpcomingMovies(movies: List<LocalMovieDto>)
-    suspend fun deleteExpiredUpcomingMovies(expirationTime: Instant, storedLanguage: String)
-}
